Import sources for 266 packages from fedora

+++ SPECS-EXTENDED/python-waitress/generate-tarball.sh
@@ -1,8 +1,14 @@
-#!/bin/sh
+#!/bin/bash
+
+if [[ $# -ne 1 ]]; then
+    echo "usage: $(basename $0) <VERSION>"
+    exit 1
+fi
 
 VERSION=$1
 
-tar -xzvf v$VERSION.tar.gz
+wget https://github.com/Pylons/waitress/archive/v$VERSION/waitress-$VERSION.tar.gz
+tar -xzvf waitress-$VERSION.tar.gz
 rm -rf waitress-$VERSION/docs
 mv waitress-$VERSION waitress-$VERSION-nodocs
-tar -czvf v$VERSION-nodocs.tar.gz waitress-$VERSION-nodocs
+tar -czvf waitress-$VERSION-nodocs.tar.gz waitress-$VERSION-nodocs
